requirements.txt
django
djangorestframework
bitly_api
flask
stripe

